==> T4064/0_T4064_preprocessing_code.ipynb <==
## 전처리를 위한 코드 쥬피터 노트북  

## import modules

import os
import pandas as pd
from pytorch_model_summary import summary
import torch

## 목표 : train.csv 파일 새로 만들기  
path, class, dir_name, gender, age, mask_state  

original_csv_path = os.path.join(os.getcwd(), 'dataset', 'train', 'train.csv') # train csv file path
image_dir_path = os.path.join(os.getcwd(), 'dataset', 'train', 'images') # train image dir path

df = pd.read_csv(original_csv_path) # read csv file
result = pd.DataFrame()
img_paths = [] # save for all train image
mask_states = [] # save for all mask state
ages = [] # save for all age
genders = [] # save for all age

for dir_name, age, gender in zip(df['path'], df['age'], df['gender']):
    each_dir_path = os.path.join(image_dir_path,dir_name)
    files_names = os.listdir(each_dir_path)

    for file_name in files_names:
        each_file_path = os.path.join(each_dir_path, file_name)
        img_paths.append(each_file_path)
        ages.append(age)
        genders.append(gender)
        if 'mask' in file_name and 'incorrect' not in file_name:
            mask_states.append('Wear')
        elif 'incorrect' in file_name:
            mask_states.append('Incorrect')
        elif 'normal' in file_name:
            mask_states.append('Not Wear')
        else:
            print(f'{file_name} is image file?')


result['path'] = img_paths
result['mask_state'] = mask_states
result['age'] = ages
result['gender'] = genders

def get_label(mask_state, gender, age):
    gender = gender.lower()
    label = 0
    if mask_state == 'Wear' and gender == 'male' and age<30: # 0:
        label = 0
    elif mask_state == 'Wear' and gender == 'male' and 30<=age<60: # 1
        label = 1
    elif mask_state == 'Wear' and gender == 'male' and age>=60: # 2
        label = 2
    elif mask_state == 'Wear' and gender == 'female' and age<30: # 3
        label = 3
    elif mask_state == 'Wear' and gender == 'female' and 30<=age<60: # 4
        label = 4
    elif mask_state == 'Wear' and gender == 'female' and age>=60: # 5
        label = 5
    elif mask_state == 'Incorrect' and gender == 'male' and age<30: # 6
        label = 6
    elif mask_state == 'Incorrect' and gender == 'male' and 30<=age<60: # 7
        label = 7
    elif mask_state == 'Incorrect' and gender == 'male' and age>=60: # 8
        label = 8
    elif mask_state == 'Incorrect' and gender == 'female' and age<30: # 9
        label = 9
    elif mask_state == 'Incorrect' and gender == 'female' and 30<=age<60: # 10
        label = 10
    elif mask_state == 'Incorrect' and gender == 'female' and age>=60: # 11
        label = 11
    elif mask_state == 'Not Wear' and gender == 'male' and age<30: # 12
        label = 12
    elif mask_state == 'Not Wear' and gender == 'male' and 30<=age<60: # 13
        label = 13
    elif mask_state == 'Not Wear' and gender == 'male' and age>=60: # 14
        label = 14
    elif mask_state == 'Not Wear' and gender == 'female' and age<30: # 15
        label = 15
    elif mask_state == 'Not Wear' and gender == 'female' and 30<=age<60: # 16
        label = 16
    elif mask_state == 'Not Wear' and gender == 'female' and age>=60: # 17
        label = 17
    else:
        raise ValueError
    return label

classes = []
for idx, row in enumerate(result.values.tolist()):
    p, s, a, g = row 
    label = get_label(s, g, a)
    classes.append(label)

result['class'] = classes

result.head(10)

new_file_root = os.path.join(os.getcwd(), 'dataset', 'train', 'custom_train.csv')
result.to_csv(new_file_root)

result['class'].value_counts()

## 목표 : Mask, Gender를 training 하기위한 csv 파일 만들기    

os.getcwd()
file_path = os.path.join(os.getcwd(), 'dataset', 'train','custom_train.csv')
target_sheet = pd.read_csv(file_path)

map_dict = {
    'Wear,male':0,
    'Wear,female':1,
    'Incorrect,male':2,
    'Incorrect,female':3,
    'Not Wear,male':4,
    'Not Wear,female':5
}

temp_list = []
for row in zip(target_sheet['mask_state'], target_sheet['gender']):
    # print(row, type(row))
    combined_row = row[0] + ',' +row[1]
    temp_list.append(combined_row)

temp_list

mapped_list = list(map(lambda x:map_dict[x],temp_list))
target_sheet['mask_gender_class'] = mapped_list

file_path = os.path.join(os.getcwd(), 'dataset', 'train','mask_gender_sheet.csv')
target_sheet.to_csv(file_path)

target_sheet

## 목표 : 3head model에 사용가능한 csv 파일 만들기  
```
dict_mask_map = {'Wear' : 0, 'Incorrect' : 1, 'Not Wear' : 2}  
dict_gender_map = {'male' : 0, 'female' : 1}  
```

df = pd.read_csv('./dataset/train/custom_train.csv')
dict_mask_map = {'Wear' : 0, 'Incorrect' : 1, 'Not Wear' : 2}  
dict_gender_map = {'male' : 0, 'female' : 1}  
df

new_df = pd.DataFrame()
new_df['path'] = df['path']
new_df['mask_class'] = df['mask_state'].map(dict_mask_map)
new_df['gender_class'] = df['gender'].map(dict_gender_map)
new_df['age'] = df['age']
new_df['origin_label'] = df['class']

new_df.to_csv('./dataset/train/3head_train.csv')

## 목표 : 모델서치

import timm

print(timm.list_models('*convnext*', pretrained=True))

model = timm.create_model('convnext_tiny',pretrained=True, num_classes=2)
print(summary(model, torch.Tensor(1,3,224,224)))

print(timm.list_models('*dense*', pretrained=True))

model = timm.create_model('densenet161', pretrained=True, num_classes=2)
print(summary(model, torch.Tensor(1,3,224,224)))
